fix build sparse

+++ src/sparse/sparse_defs.h
@@ -33,8 +33,10 @@
 #define __u16 u16
 #define __u8 u8
 
-typedef unsigned long long u64;
-typedef signed long long s64;
+//typedef unsigned long long u64;
+typedef uint64_t        u64;
+//typedef signed long long s64;
+typedef int64_t        s64;
 typedef unsigned int u32;
 typedef unsigned short int u16;
 typedef unsigned char u8;
